# src/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pod Quiz Documentation",
    "description": "API documentation for Pod Project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "http://localhost:9200"
    },
    {
      "url": "http://172.20.10.12:9200"
    }
  ],
  "paths": {
    "/api/account/register": {
      "post": {
        "tags": [
          "Web-Account"
        ],
        "summary": "Register a new account",
        "operationId": "registerAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "+1234567890"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "John"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Doe"
                  },
                  "password": {
                    "type": "string",
                    "example": "Password123!"
                  },
                  "role": {
                    "type": "string",
                    "example": "user"
                  }
                },
                "required": [
                  "phoneNumber",
                  "firstName",
                  "lastName",
                  "password",
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "200"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/restaurant/get/city": {
      "get": {
        "tags": ["Web-restaurant"],
        "summary": "Get restaurants by city and province",
        "operationId": "getRestaurantByCity",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "New York"
            }
          },
          {
            "name": "province",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "NY"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of restaurants in the specified city and province",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Restaurant": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "60d0fe4f5311236168a109cb"
                          },
                          "name": {
                            "type": "string",
                            "example": "Restaurant ABC"
                          },
                          "description": {
                            "type": "string",
                            "example": "A cozy place with great food."
                          },
                          "address": {
                            "type": "string",
                            "example": "123 Main St, City, Country"
                          },
                          "city": {
                            "type": "string",
                            "example": "New York"
                          },
                          "province": {
                            "type": "string",
                            "example": "NY"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/restaurants/{id}": {
      "get": {
        "tags": ["Web-restaurant"],

        "summary": "Get restaurant by ID",
        "operationId": "getRestaurantById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the restaurant",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Restaurant": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "60d0fe4f5311236168a109cb"
                        },
                        "name": {
                          "type": "string",
                          "example": "Restaurant ABC"
                        },
                        "description": {
                          "type": "string",
                          "example": "A cozy place with great food."
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main St, City, Country"
                        },
                        "city": {
                          "type": "string",
                          "example": "New York"
                        },
                        "province": {
                          "type": "string",
                          "example": "NY"
                        },
                        "MenuItem": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "string",
                                "example": "60d0fe4f5311236168a109cc"
                              },
                              "name": {
                                "type": "string",
                                "example": "Burger"
                              },
                              "description": {
                                "type": "string",
                                "example": "A juicy beef burger"
                              },
                              "price": {
                                "type": "number",
                                "example": 12.99
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "x-access-token",
        "in": "header"
      }
    }
  }
}